+++ NRFReceiver/NRFReceiver.cpp
@@ -5,4 +5,4 @@ int NUM_LEDS = 30;
 RF24 myRadio (7, 8);
 byte addresses[6] = {'j', 'j', 'k', 'a', 'b', 'e'};
 
-int BRIGHTNESS = 255;
+byte BRIGHTNESS = 255;
